// web/src/generated/vesting-positions/src/generated/accounts/campaign.ts
export const CAMPAIGN_DISCRIMINATOR: ReadonlyUint8Array = new Uint8Array([
  50, 40, 49, 11, 157, 220, 229, 192,
]);

export function getCampaignDiscriminatorBytes(): ReadonlyUint8Array {
  return fixEncoderSize(getBytesEncoder(), 8).encode(CAMPAIGN_DISCRIMINATOR);
}

export type Campaign = {
  discriminator: ReadonlyUint8Array;
  creator: Address;
  merkleRoot: ReadonlyUint8Array;
  start: bigint;
  end: bigint;
  cliffDuration: bigint;
  cliffReleaseBps: number;
  mintToDistribute: Address;
  isTransferable: boolean;
  gracePeriod: bigint;
  totalDeposit: bigint;
  collection: Address;
  campaignBump: number;
  collectionBump: number;
  authBump: number;
};